Rename id.ts to ReflectionSymbolId.ts
