+++ package.json
@@ -39,7 +39,7 @@
     "@types/node": "14",
     "@typescript-eslint/eslint-plugin": "^5.46.1",
     "@typescript-eslint/parser": "^5.46.1",
-    "@typestrong/fs-fixture-builder": "github:TypeStrong/fs-fixture-builder#720d17d28ba7f8e348d002d5068e728e164dff78",
+    "@typestrong/fs-fixture-builder": "github:TypeStrong/fs-fixture-builder#8abd1494280116ff5318dde2c50ad01e1663790c",
     "c8": "^7.12.0",
     "esbuild": "^0.16.8",
     "eslint": "^8.30.0",
