+++ src/lib/models/reflections/index.ts
@@ -16,4 +16,7 @@ export { SignatureReflection } from "./signature";
 export { TypeParameterReflection, VarianceModifier } from "./type-parameter";
 export { splitUnquotedString } from "./utils";
 export type { ReflectionVariant } from "./variant";
-export { ReflectionSymbolId, type ReflectionSymbolIdString } from "./id";
+export {
+    ReflectionSymbolId,
+    type ReflectionSymbolIdString,
+} from "./ReflectionSymbolId";
